chore: shadcn ui alias 변경
