+++ components.json
@@ -11,7 +11,7 @@
     "prefix": ""
   },
   "aliases": {
-    "components": "@/components",
-    "utils": "@/lib/utils"
+    "components": "@/shared/components/shadcn",
+    "utils": "@/shared/lib/tailwind/utils"
   }
 }
